+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/PeerSelection.hs
@@ -176,15 +176,15 @@ tests =
       ]
     , testGroup "bootstrap peers"
       [ testProperty "progress towards only bootstrap peers after changing to fallback state"
-                     prop_governor_only_bootstrap_peers_in_fallback_state
+                     $ prop_governor_only_bootstrap_peers_in_fallback_state . getPreGenesisMockEnv
       , testProperty "node does not learn about non trustable peers when in fallback state"
-                     prop_governor_no_non_trustable_peers_before_caught_up_state
+                     $ prop_governor_no_non_trustable_peers_before_caught_up_state . getPreGenesisMockEnv
       , testProperty "node only use bootstrap peers if in sensitive state"
-                     prop_governor_stops_using_bootstrap_peers
+                     $ prop_governor_stops_using_bootstrap_peers . getPreGenesisMockEnv
       , testProperty "node never uses non-trustable peers after clean state"
-                     prop_governor_only_bootstrap_peers_after_clean_state
+                     $ prop_governor_only_bootstrap_peers_after_clean_state . getPreGenesisMockEnv
       , testProperty "node uses ledger peers in non-sensitive mode"
-                     prop_governor_uses_ledger_peers
+                     $ prop_governor_uses_ledger_peers . getPreGenesisMockEnv
       ]
     ]
   , testGroup "issues"
@@ -1401,7 +1401,7 @@ prop_governor_target_known_2_opportunity_taken (MaxTime maxTime) env =
 
         govUseBootstrapPeersSig :: Signal UseBootstrapPeers
         govUseBootstrapPeersSig =
-          selectGovState Governor.bootstrapPeersFlag events
+          selectGovStateGenesis Governor.bootstrapPeersFlag (useGenesis env) events
 
         -- We define the governor's peer sharing opportunities at any point in time
         -- to be the governor's set of established peers, less the ones we can see
@@ -3010,6 +3010,7 @@ prop_governor_target_active_local_above (MaxTime maxTime) env =
 
 -- | When in 'TooOld' state make sure we don't stay connected to non trustable
 -- peers for too long
+
 prop_governor_only_bootstrap_peers_in_fallback_state :: GovernorMockEnvironment -> Property
 prop_governor_only_bootstrap_peers_in_fallback_state env =
     let events = Signal.eventsFromListUpToTime (Time (10 * 60 * 60))
@@ -3023,7 +3024,7 @@ prop_governor_only_bootstrap_peers_in_fallback_state env =
 
         govLedgerStateJudgement :: Signal LedgerStateJudgement
         govLedgerStateJudgement =
-          selectGovState (Governor.ledgerStateJudgement) events
+          selectGovState Governor.ledgerStateJudgement events
 
         govKnownPeers :: Signal (Set PeerAddr)
         govKnownPeers =
@@ -3071,7 +3072,7 @@ prop_governor_no_non_trustable_peers_before_caught_up_state env =
 
         govLedgerStateJudgement :: Signal LedgerStateJudgement
         govLedgerStateJudgement =
-          selectGovState (Governor.ledgerStateJudgement) events
+          selectGovState Governor.ledgerStateJudgement events
 
         govKnownPeers :: Signal (Set PeerAddr)
         govKnownPeers =
@@ -3109,6 +3110,7 @@ prop_governor_no_non_trustable_peers_before_caught_up_state env =
           Set.null
           keepNonTrustablePeersTooLong
 
+
 prop_governor_only_bootstrap_peers_after_clean_state :: GovernorMockEnvironment -> Property
 prop_governor_only_bootstrap_peers_after_clean_state env =
     let events = Signal.eventsFromListUpToTime (Time (10 * 60 * 60))
@@ -3122,7 +3124,7 @@ prop_governor_only_bootstrap_peers_after_clean_state env =
 
         govLedgerStateJudgement :: Signal LedgerStateJudgement
         govLedgerStateJudgement =
-          selectGovState (Governor.ledgerStateJudgement) events
+          selectGovState Governor.ledgerStateJudgement events
 
         govKnownPeers :: Signal (Set PeerAddr)
         govKnownPeers =
@@ -3239,7 +3241,7 @@ prop_governor_uses_ledger_peers env =
 
         govLedgerStateJudgement :: Signal LedgerStateJudgement
         govLedgerStateJudgement =
-          selectGovState (Governor.ledgerStateJudgement) events
+          selectGovState Governor.ledgerStateJudgement events
 
         govPublicRootPeersResultsSig :: Signal (PublicRootPeers PeerAddr)
         govPublicRootPeersResultsSig =
@@ -3292,7 +3294,20 @@ selectGovState :: Eq a
 selectGovState f =
     Signal.nub
   -- TODO: #3182 Rng seed should come from quickcheck.
-  . Signal.fromChangeEvents (f $! Governor.emptyPeerSelectionState (mkStdGen 42) [])
+  . Signal.fromChangeEvents (f $! Governor.emptyPeerSelectionState (mkStdGen 42) [] False)
+  . Signal.selectEvents
+      (\case GovernorDebug (TraceGovernorState _ _ st) -> Just $! f st
+             _                                         -> Nothing)
+
+selectGovStateGenesis :: Eq a
+                      => (forall peerconn. Governor.PeerSelectionState PeerAddr peerconn -> a)
+                      -> Bool
+                      -> Events TestTraceEvent
+                      -> Signal a
+selectGovStateGenesis f useGenesisFlag =
+    Signal.nub
+  -- TODO: #3182 Rng seed should come from quickcheck.
+  . Signal.fromChangeEvents (f $! Governor.emptyPeerSelectionState (mkStdGen 42) [] useGenesisFlag)
   . Signal.selectEvents
       (\case GovernorDebug (TraceGovernorState _ _ st) -> Just $! f st
              _                                         -> Nothing)
@@ -3326,8 +3341,9 @@ _governorFindingPublicRoots :: Int
                             -> STM IO LedgerStateJudgement
                             -> PeerSharing
                             -> StrictTVar IO OnlyLocalOutboundConnections
+                            -> Bool
                             -> IO Void
-_governorFindingPublicRoots targetNumberOfRootPeers readDomains readUseBootstrapPeers readLedgerStateJudgement peerSharing olocVar = do
+_governorFindingPublicRoots targetNumberOfRootPeers readDomains readUseBootstrapPeers readLedgerStateJudgement peerSharing olocVar useGenesisFlag = do
     dnsSemaphore <- newLedgerAndPublicRootDNSSemaphore
     publicRootPeersProvider
       tracer
@@ -3337,7 +3353,7 @@ _governorFindingPublicRoots targetNumberOfRootPeers readDomains readUseBootstrap
       readDomains
       (ioDNSActions LookupReqAAndAAAA) $ \requestPublicRootPeers -> do
         publicStateVar <- newTVarIO (emptyPublicPeerSelectionState @SockAddr)
-        debugVar <- newTVarIO $ emptyPeerSelectionState (mkStdGen 42) []
+        debugVar <- newTVarIO $ emptyPeerSelectionState (mkStdGen 42) [] useGenesisFlag
         countersVar <- newTVarIO $ emptyPeerSelectionCounters []
         peerSelectionGovernor
           tracer tracer tracer
@@ -3346,6 +3362,7 @@ _governorFindingPublicRoots targetNumberOfRootPeers readDomains readUseBootstrap
           countersVar
           publicStateVar
           debugVar
+          useGenesisFlag
           actions
             { requestPublicRootPeers = \_ ->
                 transformPeerSelectionAction requestPublicRootPeers }
@@ -3437,6 +3454,7 @@ prop_issue_3550 = prop_governor_target_established_below defaultMaxTime $
       pickWarmPeersToDemote = Script (PickFirst :| []),
       pickColdPeersToForget = Script (PickFirst :| []),
       peerSharing = PeerSharingEnabled,
+      useGenesis = False,
       useBootstrapPeers = Script ((DontUseBootstrapPeers, NoDelay) :| []),
       ledgerStateJudgement = Script ((YoungEnough, NoDelay) :| [])
     }
@@ -3473,6 +3491,7 @@ prop_issue_3515 = prop_governor_nolivelock $
       pickWarmPeersToDemote = Script (PickFirst :| []),
       pickColdPeersToForget = Script (PickFirst :| []),
       peerSharing = PeerSharingEnabled,
+      useGenesis = False,
       useBootstrapPeers = Script ((DontUseBootstrapPeers, NoDelay) :| []),
       ledgerStateJudgement = Script ((YoungEnough, NoDelay) :| [])
     }
@@ -3509,6 +3528,7 @@ prop_issue_3494 = prop_governor_nofail $
       pickWarmPeersToDemote = Script (PickFirst :| []),
       pickColdPeersToForget = Script (PickFirst :| []),
       peerSharing = PeerSharingEnabled,
+      useGenesis = False,
       useBootstrapPeers = Script ((DontUseBootstrapPeers, NoDelay) :| []),
       ledgerStateJudgement = Script ((YoungEnough, NoDelay) :| [])
     }
@@ -3561,6 +3581,7 @@ prop_issue_3233 = prop_governor_nolivelock $
       pickWarmPeersToDemote = Script (PickFirst :| []),
       pickColdPeersToForget = Script (PickFirst :| []),
       peerSharing = PeerSharingEnabled,
+      useGenesis = False,
       useBootstrapPeers = Script ((DontUseBootstrapPeers, NoDelay) :| []),
       ledgerStateJudgement = Script ((YoungEnough, NoDelay) :| [])
     }